+++ assets/javascripts/kitten/components/form/text-input-with-unit/index.js
@@ -75,6 +75,18 @@ const StyledInputUnit = styled.span`
       padding: 0 ${pxToRem(10)};
     `}
 
+  ${({ huge }) =>
+    huge &&
+    css`
+      padding: 0 ${pxToRem(20)};
+    `}
+
+  ${({ giant }) =>
+    giant &&
+    css`
+      padding: 0 ${pxToRem(25)};
+    `}
+
   ${({ unitWord }) =>
     unitWord &&
     css`
@@ -93,6 +105,9 @@ export class TextInputWithUnit extends PureComponent {
     valid: PropTypes.bool,
     error: PropTypes.bool,
     tiny: PropTypes.bool,
+    huge: PropTypes.bool,
+    giant: PropTypes.bool,
+    center: PropTypes.bool,
     disabled: PropTypes.bool,
     unit: PropTypes.string,
     unitWord: PropTypes.bool,
@@ -106,6 +121,9 @@ export class TextInputWithUnit extends PureComponent {
     valid: false,
     error: false,
     tiny: false,
+    huge: false,
+    giant: false,
+    center: false,
     disabled: false,
     digits: null,
   }
@@ -120,6 +138,9 @@ export class TextInputWithUnit extends PureComponent {
       valid,
       error,
       tiny,
+      huge,
+      giant,
+      center,
       disabled,
       unit,
       unitWord,
@@ -136,6 +157,9 @@ export class TextInputWithUnit extends PureComponent {
           valid={valid}
           error={error}
           tiny={tiny}
+          huge={huge}
+          giant={giant}
+          center={center}
           disabled={disabled}
           {...others}
         />
@@ -144,6 +168,8 @@ export class TextInputWithUnit extends PureComponent {
           error={error}
           disabled={disabled}
           tiny={tiny}
+          huge={huge}
+          giant={giant}
         >
           {unit || unitWord}
         </StyledInputUnit>