+++ assets/javascripts/kitten/components/form/text-input-with-unit/stories.js
@@ -26,6 +26,13 @@ const info = {
    #### Size
     ~~~js
     <TextInputWithUnit tiny />
+    <TextInputWithUnit huge />
+    <TextInputWithUnit giant />
+    ~~~
+
+   #### Alignment
+    ~~~js
+    <TextInputWithUnit center />
     ~~~
 
     #### State
@@ -55,6 +62,9 @@ storiesOf('Form/TextInput', module)
               valid={boolean('Valid', false)}
               disabled={boolean('Disabled', false)}
               tiny={boolean('Tiny', false)}
+              huge={boolean('Huge', false)}
+              giant={boolean('Giant', false)}
+              center={boolean('Center', false)}
               error={boolean('Error', false)}
               unit={text('Unit', '€')}
               digits={number('Digits', 0)}