+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `huge`, `giant` and `center` props to `TextInput`.
+- Feature: Add `huge` and `giant` prop to `SelectWithState`.
+
 ## [2.83.0] - 2020-07-31
 
 Feature: